pyproject.toml
[project]
name = "controleciclo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-docx>=1.1.2",
    "xlrd>=2.0.1",
]
